ipvs: provide iph to schedulers
Before now the schedulers needed access only to IP
addresses and it was easy to get them from mbuf.

The MH scheduler will need the protocol and ports which is
difficult to get from mbuf for the IPv6 case. As we have all the
data in the iph structure, to avoid the same slow lookups
provide the iph to schedulers.