kni: isolate kni ingress traffic using kni address flow
The idea is to specify kni interface with an ip address, and isolate
all traffic target at the address to a dedicated nic rx-queue, which
may eliminate disturbances of dataplane when overload.

Note that not all nic can support this flow type under the premise of
sapool, and an additional rx queue is required, which is configured
into dpvs's kni worker.

Signed-off-by: ywc689 <ywc689@163.com>