ipvs: expire quiescent connections after realserver was removed
This is an optimization of pr#574, in which quiescent connections
are expired immediately when realserver weight is 0. Unfortunately,
it didn't deal with the case where the realserver was removed.
The patch fixes the problem by also expiring the connections that are
in unavailable status(dp_vs_dest_is_avail returns false).

Signed-off-by: ywc689 <ywc689@163.com>