kni: fix possible alloc_q starvation
The size of kni_mbuf_pool is 65535. DPDK KNI may be starvation
when mbufs are exhausted. Many reasons cause the mbufs of
the kni_mbuf_pool to be cached, such as kni_rx_ring_*, driver.

For example, under the configuration of 16 txqs + 4096 descriptors,
since ixgbe uses lazy recycling, a maximum of 65536 mbufs will be cached.

Using a large enough mbuf pool is an easy fix. pktmbuf_pool is the best choice.

DPDK KNI starvation bug detail:
http://patches.dpdk.org/project/dpdk/patch/20221109060434.2012064-1-zhouyates@gmail.com/
http://patches.dpdk.org/project/dpdk/patch/20221230042338.1670768-1-zhouyates@gmail.com/