tools/healthcheck: fix rs weight update problem when rs's weight changed adminstratively
When healthcheck program updating the rs's weight, it should consider:

* rs's weight may change adminstratively
* rs availabilty may change due to the server is closed or crashed

Healthcheck stores the original weight when it sends down notification in order to restore
the rs's weight when it gets healthy later. The problem is original weight is syncd from
dpvs periodically and not always up-to-date, which may result in incorrect weight update in
healthcheck up notification.

Signed-off-by: ywc689 <ywc689@163.com>