tools/healthcheck: further fix the problem discussed in previous commit
Unfortunately, the CAS solution in the last commit is not effective for healthcheck
process cannot maintain the correct previous weight value. Thus resource version is
used instead, which should be passed to the lb api when the healthcheck try to update
rs, and the api would reject the request if the resouce version is outdated.

use vs version to resync target status when notification fails

Signed-off-by: ywc689 <ywc689@163.com>